+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed `Brownian` and `Red` noise generators to reset after seeking.
 - Fixed sources to correctly handle sample rate and channel count changes at span boundaries.
 - Fixed sources to detect parameter updates after mid-span seeks.
+- Fixed `Stoppable` and `Skippable` not signaling exhaustion.
 
 ## Version [0.22.2] (2026-02-22)